+++ ArtmaisBackend/Infrastructure/Repository/Interface/IContactRepository.cs
@@ -5,8 +5,8 @@ namespace ArtmaisBackend.Infrastructure.Repository.Interface
 {
     public interface IContactRepository
     {
-        Contact? GetContactByUser(int? userId);
-        Contact Create(ContactRequest contactRequest, int userId);
+        Contact? GetContactByUser(long? userId);
+        Contact Create(ContactRequest contactRequest, long userId);
         Contact Update(Contact contact);
     }
 }
